+++ app/scripts/api.js
@@ -774,7 +774,11 @@ const createApi = function api(context, pubSub) {
        *  // 2D or BEDPE-like array
        *  [["chr1", 249200621, "chr2", 50000], ["chr3", 197972430, "chr4", 50000]]
        *
-       * ``viewConfig:`` Returns the current view config.
+       * ``viewConfig:`` Returns the current view config (as a string).
+       *  This event is published upon interactions including:
+       *  - Saving in the view config editor modal.
+       *  - Panning and zooming in views, which update view object ``initialXDomain`` and ``initialYDomain`` values.
+       *  - Brushing in ``viewport-projection-`` tracks containing null ``fromViewUid`` fields, which update track object ``projectionXDomain`` and ``projectionYDomain`` values.
        *
        * ``mouseMoveZoom:`` Returns the location and data at the mouse cursor's
        * screen location.