+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed blurry exported heatmap SVG graphics issue using the `image-rendering` CSS property.
 - Enabled view-based scaling
 - Left and right tracks can now be value scale locked
+- Added the attributes `projectionXDomain` and `projectionYDomain` to the `viewport-projection-horizontal` and `viewport-projection-vertical`, respectively, and both to the `viewport-projection-center`, to support the case in which the `fromViewUid` attribute is undefined.
 - Added the `.on('createSVG')` listener to the JS API, with the corresponding `.off('createSVG')`, for manipulating exported SVGs before they are returned.
 
 _[Detailed changes since v1.8.4](https://github.com/higlass/higlass/compare/v1.8.3...develop)_